Met à jour le calcul RDCM pour le sel de dissolution
